+++ discord/client.py
@@ -30,7 +30,7 @@ import logging
 import signal
 import sys
 import traceback
-from typing import Any, Callable, Coroutine, Dict, Generator, List, Optional, Sequence, TYPE_CHECKING, Tuple, TypeVar, Union
+from typing import Any, Callable, Coroutine, Dict, Generator, List, Literal, Optional, Sequence, TYPE_CHECKING, Tuple, TypeVar, Union, overload
 
 import aiohttp
 
@@ -65,11 +65,15 @@ from .sticker import GuildSticker, StandardSticker, StickerPack, _sticker_factor
 from .application_commands import ApplicationCommand
 
 if TYPE_CHECKING:
-    from .abc import SnowflakeTime, PrivateChannel, GuildChannel, Snowflake
+    from .abc import SnowflakeTime, PrivateChannel, GuildChannel, Snowflake as DiscordObject
     from .channel import DMChannel
     from .message import Message
     from .member import Member
     from .voice_client import VoiceProtocol
+    from .reaction import Reaction
+    from .raw_models import RawReactionActionEvent
+    from .interactions import Interaction
+    from .types.snowflake import Snowflake
 
 __all__ = (
     'Client',
@@ -705,7 +709,7 @@ class Client:
         """List[:class:`~discord.User`]: Returns a list of all the users the bot can see."""
         return list(self._connection._users.values())
 
-    def get_channel(self, id: int, /) -> Optional[Union[GuildChannel, Thread, PrivateChannel]]:
+    def get_channel(self, id: Snowflake, /) -> Optional[Union[GuildChannel, Thread, PrivateChannel]]:
         """Returns a channel or thread with the given ID.
 
         Parameters
@@ -869,6 +873,106 @@ class Client:
         Waits until the client's internal cache is all ready.
         """
         await self._ready.wait()
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["message"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Message:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["reaction_add"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Tuple[Reaction, Union[User, Member]]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["raw_reaction_add"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> RawReactionActionEvent:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["reaction_remove"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Tuple[Reaction, Union[User, Member]]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["raw_reaction_remove"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> RawReactionActionEvent:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["slash_command"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Interaction[None]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["component_interaction"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Interaction[str]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["autocomplete_interaction"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Interaction[None]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["modal_submit"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Interaction[str]:
+        ...
+
+    @overload
+    async def wait_for(
+            self,
+            event: Literal["interaction"],
+            *,
+            check: Optional[Callable[..., bool]] = None,
+            timeout: Optional[float] = None,
+            ) -> Interaction:
+        ...
 
     def wait_for(
         self,
@@ -1517,7 +1621,7 @@ class Client:
         data = await self.http.list_premium_sticker_packs()
         return [StickerPack(state=self._connection, data=pack) for pack in data['sticker_packs']]
 
-    async def create_dm(self, user: Snowflake) -> DMChannel:
+    async def create_dm(self, user: DiscordObject) -> DMChannel:
         """|coro|
 
         Creates a :class:`.DMChannel` with this user.
@@ -1544,7 +1648,11 @@ class Client:
         return state.add_dm_channel(data)
 
     async def register_application_commands(
-            self, commands: List[ApplicationCommand], *, guild: Optional[Snowflake] = None) -> List[ApplicationCommand]:
+            self,
+            commands: List[ApplicationCommand],
+            *,
+            guild: Optional[DiscordObject] = None,
+            ) -> List[ApplicationCommand]:
         """|coro|
 
         Register the bot's commands as application and slash commands. Providing ``None`` or an empty list
@@ -1581,7 +1689,191 @@ class Client:
             application_info = await self.application_info()
             application_id = application_info.id
         if guild:
-            data = await self.http.bulk_upsert_guild_commands(application_id, guild.id, [c.to_json() for c in commands])
+            data = await self.http.bulk_create_guild_application_commands(application_id, guild.id, [c.to_json() for c in commands])
         else:
-            data = await self.http.bulk_upsert_global_commands(application_id, [c.to_json() for c in commands])
+            data = await self.http.bulk_create_global_application_commands(application_id, [c.to_json() for c in commands])
         return [ApplicationCommand.from_data(c) for c in data]
+
+    async def fetch_global_application_commands(self) -> List[ApplicationCommand]:
+        """
+        |coro|
+
+        Get all of the application commands that the bot has registered globally.
+
+        Returns
+        --------
+        List[:class:`ApplicationCommand`]
+            A list of added application commands.
+
+        Raises
+        -------
+        :exception:`discord.HTTPException`
+            Getting the application commands failed.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        payload = await self.http.get_global_application_commands(application_id)
+        return [
+            ApplicationCommand.from_data(i)
+            for i in payload
+        ]
+
+    async def fetch_global_application_command(self, command_id: Snowflake) -> ApplicationCommand:
+        """
+        |coro|
+
+        Get a globally added command by its ID.
+
+        Parameters
+        -----------
+        command_id: :class:`int`
+            The ID of the command that you want to get.
+
+        Returns
+        --------
+        :class:`ApplicationCommand`
+            The application command with the ID given.
+
+        Raises
+        -------
+        :exception:`discord.NotFound`
+            There is no globally added command with the ID provided.
+        :exception:`discord.HTTPException`
+            Getting the application commands failed.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        payload = await self.http.get_global_application_command(application_id, command_id)
+        return ApplicationCommand.from_data(payload)
+
+    async def create_global_application_command(self, command: ApplicationCommand) -> ApplicationCommand:
+        """
+        |coro|
+
+        Create a new application command.
+
+        Parameters
+        -----------
+        command: :class:`discord.ApplicationCommand`
+            The command that you want to add.
+
+        Returns
+        --------
+        :class:`ApplicationCommand`
+            The command that was added.
+
+        Raises
+        -------
+        :exception:`discord.HTTPException`
+            Failed to create the new command.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        payload = await self.http.create_global_application_command(application_id, command.to_json())
+        return ApplicationCommand.from_data(payload)
+
+    async def edit_global_application_command(self, command: DiscordObject, **kwargs) -> ApplicationCommand:
+        """
+        |coro|
+
+        Edit the attributes of an application command.
+
+        Parameters
+        -----------
+        command: :class:`discord.abc.Snowflake`
+            The command that you want to edit.
+        name: :class:`str`
+            The new name of the command.
+        description: :class:`str`
+            The new description of the command.
+        options: List[:class:`discord.ApplicationCommandOption`]
+            The new options of the command.
+
+        Returns
+        --------
+        :class:`discord.abc.Snowflake`
+            The command that you want to edit the attribute of.
+
+        Raises
+        -------
+        :exception:`discord.NotFound`
+            There is no globally added command with the ID provided.
+        :exception:`discord.HTTPException`
+            Getting the application commands failed.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        payload = await self.http.edit_global_application_command(application_id, command.id, kwargs)
+        return ApplicationCommand.from_data(payload)
+
+    async def delete_global_application_command(self, command: DiscordObject) -> None:
+        """
+        |coro|
+
+        Delete a global application command.
+
+        Parameters
+        -----------
+        command: :class:`discord.abc.Snowflake`
+            The command that you want to delete.
+
+        Raises
+        -------
+        :exception:`discord.NotFound`
+            There is no globally added command with the ID provided.
+        :exception:`discord.HTTPException`
+            Getting the application commands failed.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        await self.http.delete_global_application_command(application_id, command.id)
+
+    async def bulk_create_global_application_commands(self, commands: List[ApplicationCommand]) -> List[ApplicationCommand]:
+        """
+        |coro|
+
+        Set all of the global application command for the bot.
+
+        Parameters
+        -----------
+        commands: List[:class:`ApplicationCommand`]
+            A list of commands that you want to add.
+
+        Returns
+        --------
+        List[:class:`ApplicationCommand`]
+            The application commands that you added.
+
+        Raises
+        -------
+        :exception:`discord.HTTPException`
+            Getting the application commands failed.
+        """
+
+        application_id = self.application_id
+        if application_id is None:
+            application_info = await self.application_info()
+            application_id = application_info.id
+        payload = await self.http.bulk_create_global_application_commands(application_id, [
+            c.to_json()
+            for c in commands
+        ])
+        return [
+            ApplicationCommand.from_data(i)
+            for i in payload
+        ]