+++ discord/types/interactions.py
@@ -189,6 +189,7 @@ class _InteractionOptional(TypedDict, total=False):
     member: Member
     user: User
     message: Message
+    guild_locale: str
 
 
 class Interaction(_InteractionOptional):
@@ -198,6 +199,7 @@ class Interaction(_InteractionOptional):
     token: str
     version: int
     resolved: InteractionResolved
+    locale: str
 
 
 class InteractionApplicationCommandCallbackData(TypedDict, total=False):